requirements.txt
alabaster==0.7.12
astroid==2.3.3
autograd==1.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
cma==2.7.0
cycler==0.10.0
Cython==0.29.16
docutils==0.16
future==0.18.2
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.0
numpydoc==0.9.2
packaging==20.3
pandas==1.0.3
plotly==4.7.1
pyDOE2==1.3.0
Pygments==2.6.1
pylint==2.4.4
pymoo==0.4.0
pymop==0.2.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.25.8
wrapt==1.11
